fixed binary text handling in IE11
FileReader.readAsBinaryString() was deprecated (and indeed not supported
anymore by IE 11).

This reads the Blob with the standard readAsArrayBuffer() instead, and
works in all browsers AFAICT.

The rest of the code should still be updated to process the Uint8Array
directly rather than through a "binary" String, but it's an other story.